requirements.txt
streamlit
feedparser
google-generativeai
requests